refactor(deep-research): extract aggregator output check into its own module
Move aggregatorOutputComplete out of heuristic.ts into a dedicated
aggregator-output.ts helper and update the three agent workflow imports
and the test file accordingly.

Assistant-model: Claude Code
